+++ lib/pages/GuidedMode/GuidedModeUI.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:beebetter/widgets/PromptCard.dart';
+import 'package:beebetter/widgets/Cards/PromptCard/PromptCard.dart';
 import 'package:beebetter/pages/GuidedMode/GuidedModeLogic.dart';
 import 'package:flutter_card_swiper/flutter_card_swiper.dart';
