+++ lib/widgets/Cards/EntryCard/EntryCard.dart
@@ -1,54 +1,36 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:beebetter/widgets/PromptInput.dart';
-import 'package:beebetter/widgets/EmotionWheel/EmotionWheel.dart';
+import 'package:beebetter/widgets/Cards/EntryCard/EntryInput.dart';
+import 'package:beebetter/widgets/Cards/EmotionWheel/EmotionWheel.dart';
+import 'package:beebetter/pages/NewEntryPage/NewEntryPageLogic.dart';
 import 'package:flutter_flip_card/flutter_flip_card.dart';
-import 'package:flutter_card_swiper/flutter_card_swiper.dart';
-import 'package:beebetter/pages/GuidedMode/GuidedModeLogic.dart';
 
-class PromptCard extends StatefulWidget {
+class EntryCard extends StatefulWidget {
   final int index;
   final String category;
-  final String prompt;
   final bool canContinue;
-  final bool isDone;
   final void Function(String) onContinuePressed;
   final void Function(String) onTextChanged;
   final String initialText;
-  final CardSwiperController cardSwiperController;
 
-  const PromptCard({
+  const EntryCard({
     super.key,
     required this.index,
     required this.category,
-    required this.prompt,
     required this.canContinue,
-    required this.isDone,
     required this.onContinuePressed,
     required this.onTextChanged,
     required this.initialText,
-    required this.cardSwiperController,
   });
 
   @override
-  State<PromptCard> createState() => PromptCardState();
+  State<EntryCard> createState() => EntryCardState();
 }
 
-class PromptCardState extends State<PromptCard>
+class EntryCardState extends State<EntryCard>
     with AutomaticKeepAliveClientMixin {
   late TextEditingController controller;
   final cardController = FlipCardController();
-
-  // Swipe down tracking for deleting prompt
-  double dragOffset = 0.0;
-  bool showDeleteHint = false;
-
-  void resetDrag() {
-    setState(() {
-      dragOffset = 0.0;
-      showDeleteHint = false;
-    });
-  }
 
   @override
   void initState() {
@@ -57,7 +39,7 @@ class PromptCardState extends State<PromptCard>
   }
 
   @override
-  void didUpdateWidget(covariant PromptCard oldWidget) {
+  void didUpdateWidget(covariant EntryCard oldWidget) {
     super.didUpdateWidget(oldWidget);
     if (oldWidget.initialText != widget.initialText) {
       controller.text = widget.initialText;
@@ -70,8 +52,7 @@ class PromptCardState extends State<PromptCard>
   @override
   Widget build(BuildContext context) {
     super.build(context);
-    final logic = context.watch<GuidedModeLogic>();
-
+    final logic = context.watch<NewEntryPageLogic>();
     final colorScheme = Theme.of(context).colorScheme;
 
     return FlipCard(
@@ -84,11 +65,8 @@ class PromptCardState extends State<PromptCard>
         shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(12)),
         child: Padding(
           padding: const EdgeInsets.all(16.0),
-          child: PromptInput(
-            category: widget.category,
-            prompt: widget.prompt,
+          child: EntryInput(
             canContinue: widget.canContinue,
-            isDone: widget.isDone,
             controller: controller,
             onTextChanged: widget.onTextChanged,
             onContinuePressed: widget.onContinuePressed,
@@ -104,25 +82,21 @@ class PromptCardState extends State<PromptCard>
         child: EmotionWheel(
           emotionItems: logic.emotionItems,
           levels: logic.emotionLevels,
-          selectedEmotions: logic.currentPromptInfo.emotions,
+          selectedEmotions: logic.emotions,
           canSelectNext: logic.canSelectNext,
           onBack: () => cardController.flipcard(),
           onEmotionSelected: (level, emotion) {
-            logic.selectEmotion(level, emotion);
+            logic.emotions[level] = emotion;
+            logic.updateCanSelectNextForLevel(level);
           },
-          onNext: (level) async {
+          onNext: (level) {
             if (level == logic.emotionLevels - 1) {
+              logic.saveEntry();
               cardController.flipcard();
-              logic.submitEmotion(level);
-              widget.cardSwiperController.swipe(CardSwiperDirection.left);
-              await Future.delayed(const Duration(milliseconds: 300));
-              logic.submit(widget.index, widget.cardSwiperController);
-            } else {
-              logic.updateCanSelectNextForLevel(level + 1);
-            }
+            };
           },
           onLevelChanged: logic.updateCanSelectNextForLevel,
-        )
+        ),
       ),
     );
   }
