+++ lib/widgets/Cards/EntryCard/EntryInput.dart
@@ -2,9 +2,9 @@ import 'dart:math';
 
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:beebetter/pages/GuidedMode/RecordingLogic.dart';
+import 'package:beebetter/widgets/Cards/RecordingCard/RecordingLogic.dart';
 import 'package:beebetter/pages/NewEntryPage/NewEntryPageLogic.dart';
-import 'package:beebetter/widgets/RecordingCard.dart';
+import 'package:beebetter/widgets/Cards/RecordingCard/RecordingCard.dart';
 
 class EntryInput extends StatefulWidget {
   final bool canContinue;
