+++ lib/widgets/Cards/PromptCard/PromptInput.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:beebetter/pages/GuidedMode/RecordingLogic.dart';
-import 'package:beebetter/widgets/RecordingCard.dart';
+import 'package:beebetter/widgets/Cards/RecordingCard/RecordingLogic.dart';
+import 'package:beebetter/widgets/Cards/RecordingCard/RecordingCard.dart';
 import 'package:beebetter/pages/GuidedMode/GuidedModeLogic.dart';
-import 'package:beebetter/widgets/DoneCard.dart';
+import 'package:beebetter/widgets/Cards/PromptCard/DoneCard.dart';
 
 class PromptInput extends StatefulWidget {
   final String category;
