+++ lib/widgets/Cards/RecordingCard/RecordingIndicator.dart
